Failed Attempt at NativeScript.
Until Kotlin Native supports passing structs as values in a callback,
NativeScript will not be possible to achieve.

This commit adds the code that attempts to do this, but it is commented
out currently as it will not compile. Hopefully in future it will be
supported and we can revisit.