Attempt at reproducing Vector2 in Kotlin
API wise, it seems straightforward to create something nice. However it
turns out it is not possible to call interop methods that return a
struct by value. In theory a lot of the functionality could be recreated
in the Kotlin code, but the risk of divergence between Godot and Kotlin
logic is very high.